requirements.txt
google-cloud-bigquery
google-cloud-storage
pandas
scikit-learn
joblib
db-dtypes